requirements.txt
# Core Testing Framework
locust==2.43.2
pytest==9.0.2

# Monitoring & Data Persistence
influxdb==5.3.2
psutil==7.2.2

# Configuration & Data Handling
PyYAML==6.0.3
requests==2.32.5

# Utilities
schedule==1.2.2
beautifulsoup4==4.14.3
uncurl==0.0.11
python-dateutil==2.9.0.post0
